Add ref<T> class
This type makes it easy to externalize mutable observable state from components.

It is constructable either from a `var<T>` reference, a getter/setter function pair or a plain constant value, so by using `ref<T>` you've covered three use cases for the price of a single overload, which is neat. If a constant is supplied, an internal `var<T>` is managed on the heap, still allowing animation and mutation etc.

The interface mostly follows that of `var<T>`, making it easy to e.g. go from having an internal `var<T>` to externalizing the state by swapping that for `ref<T>`, or the other way around.

A notable restriction of `ref<T>` compared to `var<T>` is that you can't assign a reactive function to `ref<T>` (since that would be inconvenient to have to handle in the function pair case).